+++ lib/Cathedral/Builder/NameManager.php
@@ -43,6 +43,9 @@ class NameManager {
 	const TYPE_OTHER    = 'other';
 	/**#@-*/
 	
+	//Configuration
+	private $_config = ['entitySingular' => ['enabled' => true, 'ignore' => []]];
+	
 	protected $metadata;
 	protected $tableNames;
 	protected $tableNamesIndex;
@@ -72,6 +75,7 @@ class NameManager {
 	public $namespace_entity;
 	
 	/**
+	 * Create NameManager instance
 	 * 
 	 * @param string $namespace
 	 * @param string $tableName
@@ -94,13 +98,21 @@ class NameManager {
 		$this->setTableName($tableName);
 	}
 	
+	/**
+	 * Table to process
+	 * 
+	 * @param string $tableName
+	 */
 	public function setTableName($tableName) {
-		$this->tableName = $tableName;
-		
-		$this->init();
+		if ($tableName != null) {
+			$this->tableName = $tableName;
+			$this->init();
+		}
 	}
 	
 	/**
+	 * Namespace for the created classes
+	 * 
 	 * @param string $namespace
 	 * @throws Exception\InvalidArgumentException
 	 */
@@ -115,10 +127,13 @@ class NameManager {
 		$this->namespace_model = "{$this->namespace}\\{$this->partNameModel}";
 		$this->namespace_entity = "{$this->namespace}\\{$this->partNameEntity}";
 		
-		$this->processClassNames();
+		if (isset($this->tableName))
+			$this->processClassNames();
 	}
 	
 	/**
+	 * Array of tables in database
+	 * 
 	 * @return string[]
 	 */
 	public function getTableNames() {
@@ -126,12 +141,19 @@ class NameManager {
 	}
 	
 	/**
+	 * Table Name
+	 * 
 	 * @return string
 	 */
 	public function getTableName() {
 		return $this->tableName;
 	}
 	
+	/**
+	 * Load next table
+	 * 
+	 * @return boolean
+	 */
 	public function nextTable() {
 		if (!isset($this->tableNamesIndex)) {
 			$this->tableNamesIndex = 0;
@@ -146,39 +168,105 @@ class NameManager {
 		$this->setTableName($this->tableNames[$this->tableNamesIndex]);
 		return true;
 	}
-
-	protected function init() {
-		$this->processClassNames();
-		$this->processProperties();
+	
+	
+	/**
+	 * Enable/Disable the EntitySingular option
+	 *  Leave empty to just get current status
+	 * 
+	 * @param bool $enabled
+	 */
+	public function entitySingular($enabled = null) {
+		if (is_bool($enabled))
+			$this->_config['entitySingular']['enabled'] = $enabled;
+		
+		return $this->_config['entitySingular']['enabled']; 
 	}
 	
+	
+	/**
+	 * Return array of tables name to skip for entitySingular
+	 * 
+	 * @return array
+	 */
+	public function getEntitySingularIgnores() {
+		return $this->_config['entitySingular']['ignore'];
+	}
+	
+	
+	/**
+	 * Array of tables to ignore or string with tables delimited by pipe (|) or FALSE to clear list
+     * e.g. array('users', 'towns') or "users|towns"
+	 *  
+	 * @param array|string|false $table
+	 */
+	public function setEntitySingularIgnores($tables) {
+		if ($tables === false) {
+			$this->_config['entitySingular']['ignore'] = [];
+			$tables = [];
+		} elseif (is_string($tables)) {
+			$tables = explode('|', $tables);
+		}
+		$this->_config['entitySingular']['ignore'] = array_merge($this->_config['entitySingular']['ignore'], $tables);
+		
+		return $this;
+	}
+	
+	/**
+	 * Start processing table
+	 */
+	protected function init() {
+		if (isset($this->tableName) && (isset($this->namespace))) {
+			$this->processClassNames();
+			$this->processProperties();
+		}
+	}
+	
+	/**
+	 * Generate the related class names
+	 */
 	protected function processClassNames() {
 		$modelBaseName = ucwords($this->tableName);
 		
+		$trimWith = $this->entitySingular() ? (in_array($this->tableName, $this->getEntitySingularIgnores()) ? '' : 's') : '';
+		
 		$this->modelName			= "{$modelBaseName}Table";
-		$this->entityName			= rtrim($modelBaseName, 's');
+		$this->entityName			= rtrim($modelBaseName, $trimWith);
 		$this->entityAbstractName	= "{$this->entityName}Abstract";
-		$this->entityVariable		= rtrim($this->tableName, 's');
+		$this->entityVariable		= rtrim($this->tableName, $trimWith);
 		
 		$this->modelPath			= $this->modulePath."/{$this->partNameModel}/{$this->modelName}.php";
 		$this->entityPath			= $this->modulePath."/{$this->partNameEntity}/{$this->entityName}.php";
 		$this->entityAbstractPath	= $this->modulePath."/{$this->partNameEntity}/{$this->entityAbstractName}.php";
 	}
 	
+	/**
+	 * Generate properties
+	 * 
+	 * @throws Exception
+	 */
 	protected function processProperties() {
 		try {
 			$table = $this->metadata->getTable($this->tableName);
 		} catch (\Exception $e) {
 			throw new Exception("Table: {$this->tableName}", Exception::ERROR_DB_TABLE);
 		}
+		
 		$columns = $table->getColumns();
 		$constraints = $table->getConstraints();
+		$this->primaryIsSequence = false;
 		
 		//PRIMARY
 		foreach ($constraints AS $constraint) {
 			if ($constraint->isPrimaryKey()) {
 				$primaryColumns = $constraint->getColumns();
 				$this->primary = $primaryColumns[0];
+				
+				$sql = "SHOW COLUMNS FROM {$this->tableName} WHERE Extra = 'auto_increment' AND Field = '{$this->primary}'";
+				$stmt = \Zend\Db\TableGateway\Feature\GlobalAdapterFeature::getStaticAdapter()->query($sql);
+				$result = $stmt->execute();
+				if ($result->count())
+					$this->primaryIsSequence = true;
 			}
 		}
 		
@@ -186,21 +274,8 @@ class NameManager {
 		$this->properties = array();
 		foreach ($columns as $column) {
 			$isPrimary = false;
-			$isSequence = false;
-			if ($column->getName() == $this->primary ) {
+			if ($column->getName() == $this->primary)
 				$isPrimary = true;
-				
-				$sql = "SHOW COLUMNS FROM {$this->tableName} WHERE Field = '{$column->getName()}'";
-				$stmt = \Zend\Db\TableGateway\Feature\GlobalAdapterFeature::getStaticAdapter()->query($sql);
-				$result = $stmt->execute();
-				$row = new \ArrayObject($result->current(), \ArrayObject::ARRAY_AS_PROPS);
-				
-				if ($row->Extra == 'auto_increment') {
-					$this->primaryIsSequence = true;
-				} else {
-					$this->primaryIsSequence = false;
-				}
-			}
 			
 			$type = self::TYPE_STRING;
 			$dataType = $column->getDataType();
@@ -216,26 +291,21 @@ class NameManager {
 				$type = self::TYPE_NUMBER;
 			}
 			
-			$columnDefault = $column->getColumnDefault();
-			$default = $columnDefault;
-			if ($columnDefault == "CURRENT_TIMESTAMP") {
+			$default = $column->getColumnDefault();
+			if ($default == "CURRENT_TIMESTAMP") {
 				$default = null;
 			} elseif (strpos($dataType, 'bit') !== false) {
-				$default = (string) $columnDefault;
+				$default = (string) $default;
 				$default = (boolean) (int) $default[2];
 			}
 			
-			$this->properties[$column->getName()] = ['datatype' => $dataType, 'type' => $type, 'default' => $default, 'primary' => $isPrimary];			
+			//$this->properties[$column->getName()] = ['datatype' => $dataType, 'type' => $type, 'default' => $default, 'primary' => $isPrimary];
+			$this->properties[$column->getName()] = ['type' => $type, 'default' => $default, 'primary' => $isPrimary];
 	    }
-	    
-	    $sql = "SELECT DATABASE() AS db FROM DUAL";
-	    $stmt = \Zend\Db\TableGateway\Feature\GlobalAdapterFeature::getStaticAdapter()->query($sql);
-	    $result = $stmt->execute();
-	    $db = $result->current()['db'];
 	    
 	    # Child tables
 	    $this->relationChildren = [];
-	    $sql = "SELECT DISTINCT TABLE_NAME AS tablename FROM INFORMATION_SCHEMA.COLUMNS WHERE COLUMN_NAME = 'fk_{$this->tableName}' AND TABLE_SCHEMA='{$db}'";
+	    $sql = "SELECT DISTINCT TABLE_NAME AS tablename FROM INFORMATION_SCHEMA.COLUMNS WHERE COLUMN_NAME = 'fk_{$this->tableName}' AND TABLE_SCHEMA=(SELECT DATABASE() AS db FROM DUAL)";
 	    $stmt = \Zend\Db\TableGateway\Feature\GlobalAdapterFeature::getStaticAdapter()->query($sql);
 	    $result = $stmt->execute();
 	    while ($result->next()) {