+++ CHANGELOG.md
@@ -8,15 +8,13 @@ once done the move to history and there is my change log :)
 
 -   Features (Relations and EntitySingular) should be optional
 
--   Use field data types for better return types
-
 -   Some more convenience methods, like write [file(s) for table(s)|all]
     just simper
 
 -   Have build manager create an html table listing db\_tables and
     relates class status and view/write class
 
--   Improve code comments
+-   Improve code comments (getting there)
 
 ## Version Numbers
 
@@ -25,6 +23,24 @@ version of the project. The bit ofter the dash in the generated code
 version. I will only inclode the - if the generated code changes.
 
 ## History
+
+### 0.1.9 - 7 (2014 Jun 25)
+
+-   NameManager: improvements to SQL & speed
+
+-   NameManager: Properties reduced in size
+
+-   EntityAbstractBuilder: Code Comments
+
+-   EntityAbstractBuilder: Generate: Property types, getter & setter
+    Code comments
+
+-   Exceptions: Code Comments
+
+-   Option: EntitySingular - Enable/Disable
+
+-   Option: EntitySingular - Ignore list, if only 1 or 2 names need to
+    be skipped
 
 ### 0.1.8 - 6 (2014 Jun 17)